Avoid `promise::Error` deletion without resolving promise
The `MegaChatErrorPrivate` inherits from `promise::Error`.
The `promise::Error` can be inherited by an exception class and thrown
inside a `then()` or `fail()` handler. In the case of `MegaChatError`,
the exception will be converted to a `fail()` and the thrown
`promise::Error` exception object will be passed to a `.fail()` handler.