Stop ignoring `MSGID` when server rejects the `NEWMSG`
It may happen that the app is stopped after sending the NEWMSG but
before receiving NEWMSGID. Upon restart, the app will retry to send the
message, but if participants have changed or your own privilege doesn't
allow you to send new messages, server will reject them. Hence, we need
to move the message to the manual sending queue and ask for confirmation
before retry.