Handle disconnects done by server
When the client doesn't respond to server, the server may eventually
close the connection. The library (both libws and libwebsockets) will
notify about the event, but they won't update the state of the websocket
client (ctx).
These changes aim to solve such inconsistent connection state by (1)
deleting the libwebsockets object upon wsClose(), checking the status of
the socket before wsDisconnect() (to preven access to invalid socket)
and improving the wsIsConnected()