Skip presence's notification at termination
During termination (logout), karere disconnect(), which results on a
notification of the online presence right before the logout finishes.
The app attempts to access to the `mClient` at the
`onChatOnlineStatusUpdate()` callback, but due to the thread's
switching, it's too late and the `mClient` is already invalid.
This fix solves the issue by not notifying the invalid presence as
result of disconnect during termination.