Fix creation of backrefs
Messages used to be encrypted only once when they were added to the send
queue (and stored encrypted as blobs in the db), so the assumption was
that all messages in the send queue precede the one for which back
references are created. Now messages from the send queue are encrypted
just prior to sending, and if a send queue flush is retried, the
assumption that encrypted message is the last one is no longer true.