Persist new peers from chatd
In case a new peer join is received first from chatd than from API, the
new peer is notified but it was never saved to DB, since upon
notificatio from API, the peer was already part of the groupchat. Once
the app restarts, the peer was not anymore in the list of peers.