Do not decide locally to move message to manual-sending
The message could be already confirmed by server. It's better to send it
and receive the REJECT than to move it to the manual-sending-queue
directly, because if the server confirmed the message, we may end up
with the message in the history and the same message in the manual-
sending-queue.