Add mediainfo to attachment messages
For video files, the preview/thumbnail are set by the app and may not be
already set at the time node is checked to prepare the attachment.
However, the mediainfo (that is used to show the duration of the video)
in already available, so better to always attach the `fileattrstring`
when available.