Send ECHO independently whether MEGAchat is already in foreground
In order to quickly detect a dead-socket (closed by server while app is
paused in background in iOS), the `ECHO` is sent to server. However,
when MEGAchat already thinks it's in foreground (sending `KEEPALIVE`
instead of `KEEPALIVEAWAY`), it skips the `ECHO`.
This commit ensures the `ECHO` is always sent to detect dead-sockets
when this function is called.