Fix notification of new chatroom
This event should be notified as a `onChatListItemUpdate()` but it's
done via `onChatRoomUpdate()`, which is wrong and confuse the app when
it has another room open.
Keeping the `room` parameter as const creates a conflict, the ctor of
the `MegaChatListItem` requires to find the last text message and it
cannot be const, since it modifies the object. Since the last-text-
message may not be properly initialized when the app request a
`MegaChatListItem`, better to simply change the constness.