Support for REJECT for NEWKEY
This commit solves the situation when the client attempts to send a
`NEWKEY`+`NEWMSG`, but both are rejected. The key, however, is kept as
unconfirmed, so if later the user has privileges to send a `NEWMSG`, it
will not send a `NEWKEY`, but expects a confirmed keyid already.
We need to reset the current keyid if it's unconfirmed and got rejected.