Send a new JOIN to chatd upon re-join
When client is connected to chatd and the user is removed from a
groupchat, the opened connection will only provide updates on the
existing history (and allow to set the seen-pointer). However, if the
user re-joins (we know that because of the API actionpacket), we need to
send a new JOINRANGEHIST to chatd in order to start receiving new
history.