Do not cleanup WebRTC context until thread exit
It may happen that the karere::Client is being terminated (via a
marshall) and a request to delete the MegaChatApi instance cleans up the
rtc module, which may result on a crash if WebRTC is not completely
terminated at that point.
Instead, rtc context should be cleaned up only after all marshall calls
are executed, just before exit.