Avoid to change the state of connection in offline mode
When the app starts in offline mode, connections to chatd are kept in
state `kStateNew`. That state must be preserved until the app requests
to connect, when MEGAchat retrieves the URLs of shards.
The `ECHO` mechanism, if used during offline mode, would set the state
to `kStateDisconnected`, preventing to get the URL.