Fix loading conversations in offline mode
Android reported a bug when app is started in offline mode and the
history of a chat is fetched. MEGAchat, once all the history is loaded
in RAM and retrieved by the app, it never calls `onHistoryDone()`
because of the unexpected `nextSource == kHistSourceServerOffline`, so
the app keeps waiting forever for the corresponding
`onMessageLoaded(NULL)`.