Fix edits of items in the sending queue
Which should update the `delta` of any related message in the sending
queue, as well as its content (not only the first item in the queue, as
before).
Also, ensures the `delta` is not the same if the edit happens within the
same second than previous related item in the queue.