Fix race-condition loading contacts
When callback for fetchnodes completion is received by MEGAchat, in case
it's a new session (or there's no cache for MEGAchat), the resumption of
actionpackets must not happen until the MEGAchat client is fully
initialized. However, the `initWithNewSession()` is asynchronous and it
means that updates from actionpackets may happen before the engine is
fully initialized.
This commit solves the issue and the eventually failed assertion
`assert(mContactsLoaded)` in the callback `onChatsUpdate()`, which was
called as soon as there are pending APs received right after the
fetchnodes, while MEGAchat is fetching own keys.