Consider automatic unarchive from API in tests
It happens that API automatically unarchives the room if its archived
when chatd notifies a new message. However, if we receive a message and
very quickly archive the room, the API may find the room is archived and
unarchive it, despite the new message happened first. It's a race-
condition hard to hit, so we protect the test against it to avoid the
failure.