Make PeerChatRoom::titleString() to return C_string, not binary string
PeerChatRoom used to store the title as
`<len_firstname><firstname><lastname>`. Since it's not practical to have
a different format for the title depending on the type of chatroom, this
commit aims to keep a title as standard, but still keeps the
`Contact::titleString()` with the binary format (with the first byte
indicating the length of firstname.