Tag `ct` if created in open-mode
The `ct` in private-mode is `invitor`+[`uh1`+`key1`]*+`topic`
If the `ct` is created in open-mode, it may not be decryptable if the
user seeing the `ct` was not a participant when the `ct` was created (if
the chat is already closed). In order to solve this issue, a new TLV is
added to the `ct` in that case: `TLV_TYPE_OPENMODE = 0x10`, so the user
can recycle its unfied-key to decrypt that `ct`.