Fix load of messages after attempt while offline
When the app tries to `loadMessages()` and it requires to fetch messages
from server but we're offline, it returns `SOURCE_ERROR`. However, when
the connection is back and we are logged in into chatd, a new
`loadMessages()` should notify the received messages from server. It
loads them, but didn't notify them.
It was due to a (too) early assignment of `mServerOldHistCbEnabled` in
`getHistoryFromDbOrServer()`.
Additionally, this commit attempts to be more accurate with the
notification of `OLDMSG` after a `REJECT` of `JOINRANGEHIST` (if the
chat was not opened, we don't need to set the aforementioned flag).