Ignore DNS-lookups if outdated
Async DNS lookups do not have any timeout in libuv. In consequence,
several lookups may get queued in result to several reconnection
attempts, breaking the RetryController's logic.
In order to avoid so, a DNS-lookup will be ignored if already connected
to cache IP, if there's a newer RetryController in place (due to a
forced disconnect) or if the DNS-lookup corresponds to a
previous/outdated attempt.