Fix race condition between findLastTextMessage and getHistory
Upon `onJoinComplete()`, the `findLastTextMessage()` marshalls (1) the
decission to fetch more old messages if no last-message is available,
and that marshall call does a new marshall (2) to fetch messages (send
`HIST`). At the same time, the app may attempt to load history, which
marshalls (3) another fetch of old messages too (send `HIST`).
The resulting execution order is 1, 3, 2, and that results on the reset
of the flag `mServerOldHistCbEnabled` for the `HISTDONE` corresponding
to 3 --> the `onHistoryDone()` for 3 is not called, but expected by the
app.