Add function to handle stats by shard
Some stages need to get statistics for every shard to give more
preccision to the statistics.

We handle shard stats based on connection states. When we set a new
state, we need to get the end timestamp for the previous stage and the
start timestamp for the next one.

When the states transition indicates that an error has been happened in
a stage, we need to increment the retries number of that stage.