Do not wait for session ready in anonymous mode
In Anonymous mode there's no login/fetchnodes, so there's no need to
wait for the completion of fetchnodes in order to proceed with the
connection to chatd.
The assertion in `connect()` was failing due to the unresolved promise,
so better to keep it resolved from the scratch, which simplify the code.