Keep `ProtocolHandler` alive
It may happen that encryption of the unified key for all participants
require to fetch some keys from API. In that case, where keys are not
cached yet, the `encryptUnifiedKeyForAllParticipants()` requires API
requests and, in the meantime, the number of references to the shared
pointer `crypto` becomes 0. In result, the encryption's promise fail
(due to `throwIfDeleted()`), so does the creation of the chatroom.