Notify when `KEEPALIVE/AWAY` is actually written to socket
Ability to notify when data is written to socket, not only for sending
keepalives, but for any data, thanks to `mSendPromise`. It's created
when adding data to the output buffer and it's empty, and resolved when
the data has been written to socket. It fails if socket is not available
or anything fails.