Catchup in case chat is archived
If an archived chat receives a new message (it automatically unarchives
the room) and generates a PUSH, the client could skip the notification
because the chatroom is considered archived. In that case the client
must catchup with API to check whether the room has already been
unarchived (it should) --> generate notification.
If the room is still archived --> API issue (send event 99009)
If the PUSH is received and, right after the reception of the message,
the room is archived --> generate notification (for a message received
before the room is archived).