Avoid segfault and refactoring
If settings window was opened twice mTimeZoneDetails was NULL so exec
failed with SEGFAULT.

On the onRequestFinish (TYPE_FETCH_TIMEZONE) we need to make a copy of
MegaTimeZoneDetails.

We have created two methods to init and reset settingsWindow class, to
reuse code and avoid some errors when close the window.