Avoid to share your online status with non-contacts
Exclude any non-contact of the peerlist sent to presenced

Set and notify the Unknown status for the peers signaled with DELPEERS

If a PEERSTATUS is received for a non-contact, keep the status in
memory, but skip notification. If a non-contact becomes contact again
and there's an status in memory, notify it to the app when the user
becomes a contact again.