Update keyid in message command
It may happen that a `NEWKEY`+`NEWMSG` results in a `keyxid` being
confirmed as a definitive `keyid`, but the `msgxid` is still not
confirmed and the client reconnects to chatd. In that case it's
important to update the `keyxid`->`keyid` in the message command
(`NEWMSG`) that will be retried upon reconnection to chatd. Otherwise,
chatd will assign a `keyid=0` to the new message, which would result on
an undecryptable message.