Update gUserAttrDescs by a std::map
The array gUserAttrDescs was designed to store user attrs in the same
index position as the user attr enumeration. User attrs has been
increasing and some of them don't need to be in the cache so the attr
enum index doesn't match with the array index.

In order to fix this problem we have change the array for a std::map
which is more efficient and solve the problem because we can find by
key (user attr type)