An unknown user can be notified directly as ex-contact
The user received in the `onUsersUpdate()` is still an unknown user for
presenced's client. However, it's added to the list of users not as
contact, but directly as ex-contact. In that case, the visibility is
HIDDEN and it should not be added to the list of current peers managed
by presenced.