Fix isManagementMessage
For public chats keyid is 0 for all messages, so we can't check keyid
to identify if a message is a management message or not.

To solve this problem now we will check if the userid is the API user,
except for truncate messages and if message type is a management known
type.