Avoid race conditions on DelegateChatRoomListener
Callbacks from MegaChatRoomListener are queued by
DelegateChatRoomListener in order to be executed by the app's thread. In
result, the execution may happen when the chatroom has been closed.

Since the MegaChatRoomListenerInterface is reused by the app for a new
DelegateChatRoomListener when another chatroom is opened, it may receive
outdated callbacks, corresponding to the already closed chatroom.

This commit aims to avoid such race condition by properly invalidating
the listener when the chatroom is closed.