Keep pending edits upon reconnect
When an edited message is edited again, but being offline, upon
reconnection chatd sends the previous edit, which used to discard the
pending (newer) edit. That edition should be resent if it is newer than
the received edition.
Additionally, if a message is truncated, any further edit is futile, so
the item can be discarded from the pending queue as well.