Simplify usage of marshall calls
The `appCtx` is used everywhere in MEGAchat, since it's required to use
`marshallCall()`. I makes it easier if the `appCtx` is not passed all
the way to wherever it's used, but initialized at startup and used at
exclusively where it's mandatory, inside the `marshallCall()` in order
to post the message.