Remove not needed GroupChatRoom::mPublicChat
Better to base any decision on the chat-mode indicated by strongvelope,
which maintains whether it is a public or private chat.

The `Member` ctor needs to know if user-attrs should be fetched or not
(depending on the mode of the chat). Hence, the mode is passed as param
(`Member` ctor cannot access to strongvelope when called from
`GroupChatRoom` ctor, since it's not instantiated yet).

Also, the `addMember()` will be used only for cases when the member is
added, not by the `GroupChatRoom` ctors.