Code clean up
I've just tested the truncate case. MEGAchat intermediate layer already
checks the `msgid` exists in the given chatid, the SDK's check whether
the `msgid` is not UNDEF. Sending a random `msgid` to API results on
success on API (it doesn't check the id), but chatd does not post any
new message.
In consequence, truncates must be edits of existing messages at all
times.