Set a minimum timeout to reduce the number of retention history checks
If we send a bunch of messages in a small time period, and then we set
a retention time for the chat; when the retention time expires for any
message, the next timer will be set until next one expires and so on.

If messages were sent in a small time period a lot of checks will be
done.