Fetch fresh URL if wsConnect fails
Upon wsConnect fails we need to re-fetch a fresh URL, because we can't
detect if connection has been refused by server or we don't have
connectivity.

If host of fresh URL does not match with the host that we have stored
in cache, we need to abort the retry controller if any, update DNS
cache and retry a new connection attempt.