Select right key for reaction in private mode
Upon encrypt/decrypt a reaction of a message, the key to encrypt the
reaction depends on the key used to encrypt the message. Hence, when
reacting to a message posted in EKR off, but the chat is EKR on, the key
should be the unified key, not the specific key of the message (which
does not exist).