Avoid commit transaction, if there's not an opened transaction in db
We have seen that in some cases mHasOpenTransaction is true upon
execution of commitTransaction, however by executing query "COMMIT
TRANSACTION" an exception is thrown.

In order to detect this issue, we are going to check if DB is in
autocommit mode or not, before executing the query.

Autocommit mode is disabled by a BEGIN statement, which means that
there's an opened transaction.