Remove call to strerror_r(...).
The rationale being that the current use of this function requires us to
mess with _GNU_SOURCE which results in some pretty major problems when
building for Linux.

The most serious of these problems being that several pthreads function
are not exported unless _GNU_SOURCE is set and unfortunately enough for
us, the C++ threads library requires those.