Add CommandsQueue for SfuConnection
This queue ensures that all commands sent to SFU from client, will be
processed and sent sequentially, as SFU does not support multiple
commands per websockets message.

The queue must be accessed from a single thread (it's mandatory). In
case we need to access to it from another thread, we would need to
implement a synchronization mechanism (like a mutex)