Avoid to connect() from app's thread
With the automatic connection started by MEGAchat without the need of
the app to call `connect()`, the LWS `wsConnect()` needs to happen in
the MEGAchat's main thread. If done by the app's thread (inside the call
to `initLeanMode()`, used by the NSE), it will fail at assert when
checking the thread.