Modify SDP to enable SVC
By default SVC is not enabled, but we can enable with
webrtc::field_trial::InitFieldTrialsFromString, however we observed
that with this solution, we didn't had more that one encoding item in
transceiver parameters.

There's another workaround called SDP munging that consist on manually
modifying SDP message generated with createOffer before provide to
setLocalDescription.