Set kInitialvthumbCount to zero
In order to follow the same behavior that webclient, we will set
kInitialvthumbCount to zero, in order to not sending vthumbs
param in JOIN command.

With this change we won't request any vthumb track when we JOIN a call,
and apps will be responsible to request those tracks they need.