Add DNS cache method for records based on host
Currently chatd/presenced records can be uniquely identified by it's
shard number.

However SFU servers are not distributed by shards so we need to
identify those records by host. On the other hand we don't need to
store URL path as we will get a fresh SFU URL upon each attemp of
creating/joining a call.

We also have to comment that as multiple calls are allowed and more
than one could share the same host and being reconnecting at the same
time, some collisions could be produced on DNS cache trying to update
the same record, but It will only means an additional connection
attempt.

Finally there's a corner case to have into account. As number of SFU
records is limited, in case we reach the maximum, mCurrentShardForSfu
will be reset, and oldest records will be overwritten by new ones.
If this has happened when we try to update IP's for a particular host
that has been removed while DNS resolution finished, we only have to
add it again to fix the problem.