Enforce DNSrecord ctor
Some code paths used a DNS record that whose member were not initialized
(only the host), which causes problems when restoring from cache (the
TLS session won't be reused, since port doesn't match).
This commit aims to avoid this bug in the future, but ensuring that all
DNS records are created with URL or Host+Session (and both will set the
port).