Avoid calling onSocketClose if state is already disconnected
Calling onSocketClose more than once when state is already
disconnected, could affect to retry controller logic making that
connection is never establish.

In particular rejecting mConnectPromise could generate an unexpected
behavior when we reach connected state and we resolve the promise.

If we have rejected that promise as result of a wrong call to
onSocketClose, we could not execute lambda in then handler.