Send event to stats if there's no retry controller upon onSocketClose(kDisconnected)
This is not an expected scenario, with changes introduced in ticket
CHT-573 we will create a new retry controller instance if this happens,
but It's important to identify the codepaths where it happens, in order
to review if logic is right.

As we don't know if this case is just theorical or it could be
reproduced, it's better to send a event to stats server in order to
check it.