Revert wrong adjustments in SfuConnection related to SfuDataReception test
Previously on test SfuDataReception we created an instance of
SfuConnection with a NULL ctx, so we added a protection in
SfuConnection to avoid cancel timer if there wasn't a valid ctx.

This is a bad decision, as we shouln't modify production code just to
avoid an error in a test.

The expected behavior for SfuDataReception is to check if received data
from SFU has the right format and if it could be processed properly
depending on the received command type.

SfuDataReception test has been adjusted to avoid this unnecesary
checkups so we remove them on this commit.