Fix an overwritten app-context for several clients in 1 process
When instantiating several MegaChatImpl the application context
provided for several was being overwritten on each RTC init call.

This scenario was happening in the automatic test introduced in
CHT-636.

The fix consists in using local member variables on each instance to
keep the proper application context (i.e. MegaChatImpl) pointer.