Avoid to reset encryption halted upon join
When a new message is about to be encrypted and it requires a new key,
and the latter requires to retrieve data/keys from API, it may happen
that MEGAchat completes the JOIN to the chatid, but it should NOT
attempt to encrypt the message again. Instead, simply wait for the API's
response. It will unlock the output sending queue.