Check own userhandle upon call ctor
Under some conditions there was a crash on call ctor probably because
user handle returned by getMyUserHandle was null. It only should happen
if MegaApi object is not logged in.

In case this happens we need to avoid crash, send an event to stats
server and make an assert fail.