Adjust upper and lower packet loss limits values
Some networks have constant packet loss and it's not just 1-2 per
second but may reach ten or so. And this is not necessarily an
indication of congested network.

In such cases, the current implementation will drop the video quality
to the worst.

We nees to adjust mPacketLostLower and mPacketLostCapping to a very
high level.

Add capping of the averaged packet lost to limit the influence of a
large momentary peak on the moving average.