Parse packetsSent and totalPacketSendDelay from RTCStatsReport
packetsSent represents the total number of packets sent over this
transport.

totalPacketSendDelay is the total number of seconds that packets have
spent buffered locally before being transmitted onto the network. The
time is measured from when a packet is emitted from the RTP packetizer
until it is handed over to the OS network socket. This measurement is
added to totalPacketSendDelay when packetsSent is incremented.