Replace #ifdef by #if directive
TARGET_OS_IPHONE is always defined (for builds on mac), but its value
is either 0 or 1.  So #ifdef always evaluates true, whereas #if
evaluates to the 0 or 1 (false or true) according to the value
TARGET_OS_IPHONE is defined to be.

We need to replace all #ifdef TARGET_OS_IPHONE occurrences by #if
TARGET_OS_IPHONE on develop.