Fix race conditions on lastError member for MEGAChat tests
A RequestTracker struct has been added in order to create specific
listeners for every command to SDK. The code has been ported from
SDK's tests with an additional constructor receiving a listener to
forward the onFinishRequest callback.

SDK and MEGAChat tests should be refactored as well in order to have a
common library with all utilities like this RequestTracker.

This fix involves only the usage of lastError (which is removed) and a
further refactor of the rest of similar member variables is required.