Add orderedRemoveCall method
When we want to remove call we need to differentiate between two
scenarios. The first one is when we need to perform an ordered
disconnection from SFU (BYE command), and the second one is when we
just need to disconnect from media channel (if needed) and remove call.

In case we are sending bye command, if another action triggers a
removeCall when BYE command is delivered and we call onSendByeCommand
that call could be already destroyed.