Check if call has been destroyed if we haven't received it
Test signalling calls fails in the following use case
1) B logouts
2) A calls B (1on1 chat)
3) B logins
4) B rejects the call

The problem is that after complete login, B doesn't receive the call,
because the login process took so many time, and call was destroyed
with kNoAnswer endCallReason.

In case we don't receive the call for account B, we need to check if we
have a received a TYPE_CALL_ENDED management message, indicating that
call has been destroyed.

The main reason behind this, is differentiate the legitimate scenarios
where login takes so many time which implies that call will finish with
kNoAnswer, and those scenarios where a CALLSTATE is not received due to
an issue on chatd or SFU.