Fix unread count upon new chat JOIN
When we JOIN a new chat we request last 32 messages (-32).
We currently incremented mLastServerHistFetchCount when we receive an
old message from server, and upon HISTDONE we just set mHaveAllHistory
if we have received zero messages from server.

We also need to add a new condition to check if we have received less
messages than requested, and in that case set mHaveAllHistory true.