Check mSpeakPermission upon removeSpeaker/addSpeaker
If we receive a mid for a peer in ANSWER or AV command mSpeakPermission
must be true, otherwise SFU is sending mid for a peer without speak
request or there's a logic error on karere.

If we receive a AV command with audio flag disabled and we have an
audio slot associated to that session mSpeakPermission must be true,
otherwise there's a logic error on karere.