Set enableAudio as false by default
With protocol V2 or greater SFU doesn't allow to receive audio flag
enabled at JOIN command, so upon start or answering a new call we need
to ensure that audio flag is never enabled.

In case speak request is enabled, user need to send AV command to
unmute

In case speak request is disabled but user wants to start/answers call
with audio enabled, apps will call MegaCHatApi methods with enableAudio
== true. Internally karere will send audio flag disabled upon JOIN
command and just after that will send an AV command to enable audio.

This is a SFU requirement.