Adjust conditions to reuse video track
When we request high res video, we can instruct SFU to reuse same track
if we previously was receiving low res.

We also need to check if peer is sending, video from camera and screen
share simultaneosly. In this case we cannot reuse track as it doesn't
have sense.

Track reusing is only allowed if we are receiving 1 video track (Camera
or Screen share) in low resolution, and we want to start receiving the
same track in high resolution.