Call onUserLeave for own user in preview mode
When chatd send us a JOIN command with priv -1 and user AAAAAAAAAAA, it
means that we are in preview mode in this chat and our privilege has
been updated to not present (-1), probably because chatlink has been
invalidated.

Add sanity checks to ensure that we don't receive user AAAAAAAAAAA, for
a chat where we are not in preview mode.