Avoid race conditions when slot needs to be reassigned
In case we need to attach slot to a session (still pending to verify)
and then we need to reassign the same slot to another session (also
unverified yet) we can get a race condition.

To avoid it, we check that Cid we want to attach to the session,
matches with Cid once PeerVerificationPms is resolved.