Send 'mcru' along with OP_RINGUSER
There's a bug on the Ring user functionality for an ongoing call, as it
seemed that chatd was sending a message to the client saying there is
an incoming call, but it appeared that chatd was not telling the API to
send the corresponding push notification. This means that the call
would only ring if the user had an existing chatd connection.

To fix this issue we can send a 'mcru' command along with OP_RINGUSER

This change will need to be removed when chatd adapt it's logic to
manage internally