Set frame source upon frameComplete call
Frames from remote, local camera and local screen are received in the
same callback OnFrame but implementes on different classes. In all of
them we finally will call to frameComplete method where we will lose
information about the source.

With these changes we will provide sourceType when getImageBuffer is
called.