Move mRaiseHands clear at call::clearResources
We need to clear mRaiseHands vector if termcode is kUserHangup, so If
we join again to the same call later, we should not send JOIN with rh
flag.

However if we clear just before we send Bye command we may receive a
RHAND_DEL by SFU of another user. As mRaiseHands has been previously
cleared we won't find that user so an assert will fail.

With this change we first send BYE command and we will clear vector
upon SFU response or in case we loose connectivity with SFU.