Replace outdated staging SFU id in automated tests
Disable RaiseHandToSpeakSfuV3

RaiseHandToSpeak feature (not lite version that is working) is still
unimplemented on apps, so we don't need to test this for this moment.

This feature probably will need to be re-checked and modified before
apps implements it so that will be a good moment to adjust this test
and re enable.